+++ projects/angular-cesium/src/lib/angular-cesium-widgets/components/circles-editor/circles-editor.component.ts
@@ -5,7 +5,7 @@ import { EditActions } from '../../models/edit-actions.enum';
 import { AcLayerComponent } from '../../../angular-cesium/components/ac-layer/ac-layer.component';
 import { CoordinateConverter } from '../../../angular-cesium/services/coordinate-converter/coordinate-converter.service';
 import { MapEventsManagerService } from '../../../angular-cesium/services/map-events-mananger/map-events-manager';
-import { Subject } from 'rxjs';
+import { Subject, Subscription } from 'rxjs';
 import { CameraService } from '../../../angular-cesium/services/camera/camera.service';
 import { EditPoint } from '../../models/edit-point';
 import { CirclesManagerService } from '../../services/entity-editors/circles-editor/circles-manager.service';
@@ -102,6 +102,7 @@ import { EditableCircle } from '../../models/editable-circle';
 })
 export class CirclesEditorComponent implements OnDestroy {
   private editLabelsRenderFn: (update: CircleEditUpdate, labels: LabelProps[]) => LabelProps[];
+  private editorUpdatesSubscription: Subscription;
   public Cesium = Cesium;
   public editPoints$ = new Subject<AcNotification>();
   public editCircles$ = new Subject<AcNotification>();
@@ -123,7 +124,7 @@ export class CirclesEditorComponent implements OnDestroy {
   }
 
   private startListeningToEditorUpdates() {
-    this.circlesEditor.onUpdate().subscribe(update => {
+    this.editorUpdatesSubscription = this.circlesEditor.onUpdate().subscribe(update => {
       if (update.editMode === EditModes.CREATE || update.editMode === EditModes.CREATE_OR_EDIT) {
         this.handleCreateUpdates(update);
       } else if (update.editMode === EditModes.EDIT) {
@@ -287,6 +288,7 @@ export class CirclesEditorComponent implements OnDestroy {
   }
 
   ngOnDestroy(): void {
+    this.editorUpdatesSubscription.unsubscribe();
     this.circlesManager.clear();
   }