+++ projects/angular-cesium/src/lib/angular-cesium-widgets/components/polylines-editor/polylines-editor.component.ts
@@ -6,7 +6,7 @@ import { EditActions } from '../../models/edit-actions.enum';
 import { AcLayerComponent } from '../../../angular-cesium/components/ac-layer/ac-layer.component';
 import { CoordinateConverter } from '../../../angular-cesium/services/coordinate-converter/coordinate-converter.service';
 import { MapEventsManagerService } from '../../../angular-cesium/services/map-events-mananger/map-events-manager';
-import { Subject } from 'rxjs';
+import { Subject, Subscription } from 'rxjs';
 import { CameraService } from '../../../angular-cesium/services/camera/camera.service';
 import { EditPoint } from '../../models/edit-point';
 import { PolylinesEditorService } from '../../services/entity-editors/polyline-editor/polylines-editor.service';
@@ -84,6 +84,7 @@ import { LabelProps } from '../../models/label-props';
 })
 export class PolylinesEditorComponent implements OnDestroy {
   private editLabelsRenderFn: (update: PolylineEditUpdate, labels: LabelProps[]) => LabelProps[];
+  private editorUpdatesSubscription: Subscription;
   public Cesium = Cesium;
   public editPoints$ = new Subject<AcNotification>();
   public editPolylines$ = new Subject<AcNotification>();
@@ -106,7 +107,7 @@ export class PolylinesEditorComponent implements OnDestroy {
   }
 
   private startListeningToEditorUpdates() {
-    this.polylinesEditor.onUpdate().subscribe((update: PolylineEditUpdate) => {
+    this.editorUpdatesSubscription = this.polylinesEditor.onUpdate().subscribe((update: PolylineEditUpdate) => {
       if (update.editMode === EditModes.CREATE || update.editMode === EditModes.CREATE_OR_EDIT) {
         this.handleCreateUpdates(update);
       } else if (update.editMode === EditModes.EDIT) {
@@ -150,6 +151,7 @@ export class PolylinesEditorComponent implements OnDestroy {
           this.editPointsLayer,
           this.editPolylinesLayer,
           this.coordinateConverter,
+          this.cesiumService.getScene(),
           update.polylineOptions,
         );
         break;
@@ -218,6 +220,7 @@ export class PolylinesEditorComponent implements OnDestroy {
           this.editPointsLayer,
           this.editPolylinesLayer,
           this.coordinateConverter,
+          this.cesiumService.getScene(),
           update.polylineOptions,
           update.positions,
         );
@@ -291,6 +294,7 @@ export class PolylinesEditorComponent implements OnDestroy {
   }
 
   ngOnDestroy(): void {
+    this.editorUpdatesSubscription.unsubscribe();
     this.polylinesManager.clear();
   }